DREEM-104070670: Got a working implementation of dynamic editing completed. There is still room for improvement, but I think it does what is necessary. It turns out I did not need to implement not changing the file on disk. Instead I save a "preview" copy of a composition periodically and the clients that want "dynamic" behavior hit that url which is the same as the compostion with /preview/ prefixed to the path.
Added support to GlobalIdle to detect when the browser window is shown/hidden. Used the new onenvironmenthidden event to reset down keys in GlobalKeys so that keys do not get stuck in the on state. Added functionality to the editor to launch a "preview" tab. The tab is launched by option clicking on the existing "preview" button.

Add support for a preview=1 query param that will redirect to the /preview/... url