DREEM-104070679. Initial fix for adding dynamic editing support for Dali.
I found that dynamic editing works if the dali client is started after the application is already in the editor. When you start dali, you add the -preview flag to use preview mode. Unfortunately, when the application starts running in dali, it shows the original state of the application, not the current edit state. However, when you modify an element, it will change in dali.

I've also seen the dali client crash when new elements are added to the screen. For example, when adding a new text or bitmap item, the dali client crashes.