Fix path when requiring DALi Node.js binary on Tizen
On Tizen we need to use the path /usr/lib/node_modules/npm/node_modules/dali/dali when requiring DALi binary from Node.js. This change will test if we are on Tizen by looking for the /etc/tizen-release file. If the file is found, the Tizen path is used for the require.